Pabeidzu 13

+++ controllers/posts/delete.php
@@ -1,9 +1,12 @@
 <?php 
-
-    $sql = 'DELETE FROM posts WHERE id = :id';
-
-    $params = ["id" => $_POST["id"]];
-    $post = $db->query($sql, $params)->fetch();
-    header("Location: /");
-    exit();
+    if($_SERVER["REQUEST_METHOD"] == "POST"){
+        $sql = 'DELETE FROM posts WHERE id = :id';
+        $params = ["id" => $_POST["id"]];
+        $post = $db->query($sql, $params)->fetch();
+        header("Location: /"); 
+        exit();
+    }
+    else{
+        redirectIfNotFound();
+    }
 ?>
